requirements.txt
argparse
pyyaml
python-dotenv
colorama